Fix unsafe narrowing

+++ src/Algorithms/AlgorithmCiftiSeparate.cxx
@@ -654,7 +654,7 @@ void AlgorithmCiftiSeparate::getCroppedVolSpace(const CiftiFile* ciftiIn, const 
     int64_t numVoxels = (int64_t)myMap.size();
     if (numVoxels > 0)
     {//make a voxel bounding box to minimize memory usage
-        int extrema[6] = { myMap[0].m_ijk[0],
+        int64_t extrema[6] = { myMap[0].m_ijk[0],
             myMap[0].m_ijk[0],
             myMap[0].m_ijk[1],
             myMap[0].m_ijk[1],
@@ -699,7 +699,7 @@ void AlgorithmCiftiSeparate::getCroppedVolSpaceAll(const CiftiFile* ciftiIn, con
     int64_t numVoxels = (int64_t)myMap.size();
     if (numVoxels > 0)
     {//make a voxel bounding box to minimize memory usage
-        int extrema[6] = { myMap[0].m_ijk[0],
+        int64_t extrema[6] = { myMap[0].m_ijk[0],
             myMap[0].m_ijk[0],
             myMap[0].m_ijk[1],
             myMap[0].m_ijk[1],
